requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9
pydantic>=2.6
sqlalchemy>=2.0
aiosqlite>=0.19
python-dotenv>=1.0

# Optional providers (install as needed)
openai>=1.30.0
anthropic>=0.30.0
google-generativeai>=0.7.2
httpx>=0.27

# Optional local LLM
llama-cpp-python>=0.2.90
